lib,zebra: support virtual interface index
Add new interface field to hold the virtual interface index and
propagate it to all daemons. The virtual interface index maps one or
more real interface index to an OS virtual interface (so you can use it
to pick outgoing interface in IPv6).

This is not used when southbound doesn't provide a VIF.

Signed-off-by: Rafael Zalamena <rzalamena@opensourcerouting.org>