bgpd: Revalidate locally originated routes also when RPKI state changes
If we have something like:

```
router bgp 65001
 redistribute connected route-map rpki
 network 10.0.0.1/32 route-map rpki
 ...
!
route-map rpki permit 10
 match rpki valid
 set local-preference 150
route-map rpki permit 20
 match rpki notfound
 set local-preference 50
!
```

Then connected/static routes are never revalidated when RPKI state changes.
E.g. if it was valid, and moves to notfound => local-preference remains 150,
but should be 50.

With this patch we force redistributed/static routes to be revalidated when
revalidation event is triggered from RPKI module.

Signed-off-by: Donatas Abraitis <donatas@opensourcerouting.org>