lldpd: Code dump
(s. lldpd/README.md)

This is some code i found in some abandoned GitHub repo;
https://github.com/zhurish/route-platform

It seemed to be mostly GPLv2 so i stole it and put it here :-)

At least it will serve as a starting point to get an LLDP implementation into
frr that does not suck as hard in terms of configuration as the one distributed
with most operating systems these days.

The code has been reformatted with clang-format and some files have been
renamed.

It does not compile. It might in the future.

Signed-off-by: David Schweizer <dschweizer@opensourcerouting.org>